cmake as project manager instead of makefile

+++ shell/clean.sh
@@ -2,7 +2,9 @@
 
 cd ../server/bin
 
-rm log/*
-rm error
-rm printf
-rm core*
+[ -e log ] && rm log/*
+[ -e error ] && rm error
+[ -e printf ] && rm printf
+rm core* 2> /dev/null
+
+echo "clean done"
